% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_workspace.R
\name{count}
\alias{count}
\title{Count the number of objects inside a workspace or multiprocessing}
\usage{
count(x)
}
\arguments{
\item{x}{the \code{workspace} or the \code{multiprocessing}.}
}
\description{
Generic functions to count the number of \code{multiprocessing} (respectively \code{sa_item})
inside a \code{workspace} (respectively \code{multiprocessing}).
}
\examples{
wk <- new_workspace()
mp <- new_multiprocessing(wk, "sap1")
count(wk) # 1 multiprocessing inside the workspace wk
count(mp) # 0 sa_item inside the multiprocessing mp


}
\seealso{
Other functions to retrieve information from a workspace, multiprocessing or sa_item: \code{\link{get_model}}, \code{\link{get_name}}, \code{\link{get_ts}}.
}
