% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{x13_spec_def}
\alias{x13_def}
\alias{tramoseats_spec_def}
\alias{tramoseats_def}
\alias{regarima_spec_def_tramoseats}
\alias{regarima_def_tramoseats}
\alias{regarima_def_x13}
\alias{regarima_spec_def_x13}
\title{Deprecated functions}
\usage{
x13_spec_def(spec = c("RSA5c", "RSA0", "RSA1", "RSA2c", "RSA3", "RSA4c",
  "X11"), estimate.from = NA_character_, estimate.to = NA_character_,
  estimate.first = NA_integer_, estimate.last = NA_integer_,
  estimate.exclFirst = NA_integer_, estimate.exclLast = NA_integer_,
  estimate.tol = NA_integer_, transform.function = c(NA, "Auto",
  "None", "Log"), transform.adjust = c(NA, "None", "LeapYear",
  "LengthOfPeriod"), transform.aicdiff = NA_integer_,
  usrdef.outliersEnabled = NA, usrdef.outliersType = NA,
  usrdef.outliersDate = NA, usrdef.outliersCoef = NA,
  usrdef.varEnabled = NA, usrdef.var = NA, usrdef.varType = NA,
  usrdef.varCoef = NA, tradingdays.option = c(NA, "TradingDays",
  "WorkingDays", "UserDefined", "None"), tradingdays.autoadjust = NA,
  tradingdays.leapyear = c(NA, "LeapYear", "LengthOfPeriod", "None"),
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA, "Remove",
  "Add", "None"), easter.enabled = NA, easter.julian = NA,
  easter.duration = NA_integer_, easter.test = c(NA, "Add", "Remove",
  "None"), outlier.enabled = NA, outlier.from = NA_character_,
  outlier.to = NA_character_, outlier.first = NA_integer_,
  outlier.last = NA_integer_, outlier.exclFirst = NA_integer_,
  outlier.exclLast = NA_integer_, outlier.ao = NA, outlier.tc = NA,
  outlier.ls = NA, outlier.so = NA, outlier.usedefcv = NA,
  outlier.cv = NA_integer_, outlier.method = c(NA, "AddOne", "AddAll"),
  outlier.tcrate = NA_integer_, automdl.enabled = NA,
  automdl.acceptdefault = NA, automdl.cancel = NA_integer_,
  automdl.ub1 = NA_integer_, automdl.ub2 = NA_integer_,
  automdl.mixed = NA, automdl.balanced = NA,
  automdl.armalimit = NA_integer_, automdl.reducecv = NA_integer_,
  automdl.ljungboxlimit = NA_integer_, automdl.ubfinal = NA_integer_,
  arima.mu = NA, arima.p = NA_integer_, arima.d = NA_integer_,
  arima.q = NA_integer_, arima.bp = NA_integer_,
  arima.bd = NA_integer_, arima.bq = NA_integer_,
  arima.coefEnabled = NA, arima.coef = NA, arima.coefType = NA,
  fcst.horizon = NA_integer_, x11.mode = c(NA, "Undefined", "Additive",
  "Multiplicative", "LogAdditive", "PseudoAdditive"),
  x11.seasonalComp = NA, x11.lsigma = NA_integer_,
  x11.usigma = NA_integer_, x11.trendAuto = NA,
  x11.trendma = NA_integer_, x11.seasonalma = NA_character_,
  x11.fcasts = NA_integer_, x11.bcasts = NA_integer_,
  x11.excludeFcasts = NA)

x13_def(series, spec = c("RSA5c", "RSA0", "RSA1", "RSA2c", "RSA3",
  "RSA4c"), userdefined = NULL)

tramoseats_spec_def(spec = c("RSAfull", "RSA0", "RSA1", "RSA2", "RSA3",
  "RSA4", "RSA5"), estimate.from = NA_character_,
  estimate.to = NA_character_, estimate.first = NA_integer_,
  estimate.last = NA_integer_, estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_, estimate.tol = NA_integer_,
  estimate.eml = NA, estimate.urfinal = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.fct = NA_integer_, usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA, usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA, usrdef.varEnabled = NA, usrdef.var = NA,
  usrdef.varType = NA, usrdef.varCoef = NA, tradingdays.mauto = c(NA,
  "Unused", "FTest", "WaldTest"), tradingdays.pftd = NA_integer_,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined",
  "None"), tradingdays.leapyear = NA,
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA,
  "Separate_T", "Joint_F", "None"), easter.type = c(NA, "Unused",
  "Standard", "IncludeEaster", "IncludeEasterMonday"),
  easter.julian = NA, easter.duration = NA_integer_,
  easter.test = NA, outlier.enabled = NA,
  outlier.from = NA_character_, outlier.to = NA_character_,
  outlier.first = NA_integer_, outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_, outlier.exclLast = NA_integer_,
  outlier.ao = NA, outlier.tc = NA, outlier.ls = NA,
  outlier.so = NA, outlier.usedefcv = NA, outlier.cv = NA_integer_,
  outlier.eml = NA, outlier.tcrate = NA_integer_,
  automdl.enabled = NA, automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_, automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_, automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_, automdl.ljungboxlimit = NA_integer_,
  automdl.compare = NA, arima.mu = NA, arima.p = NA_integer_,
  arima.d = NA_integer_, arima.q = NA_integer_,
  arima.bp = NA_integer_, arima.bd = NA_integer_,
  arima.bq = NA_integer_, arima.coefEnabled = NA, arima.coef = NA,
  arima.coefType = NA, fcst.horizon = NA_integer_,
  seats.approx = c(NA, "None", "Legacy", "Noisy"),
  seats.trendBoundary = NA_integer_, seats.seasdBoundary = NA_integer_,
  seats.seasdBoundary1 = NA_integer_, seats.seasTol = NA_integer_,
  seats.maBoundary = NA_integer_, seats.method = c(NA, "Burman",
  "KalmanSmoother", "McElroyMatrix"))

tramoseats_def(series, spec = c("RSAfull", "RSA0", "RSA1", "RSA2", "RSA",
  "RSA4", "RSA5"), userdefined = NULL)

regarima_spec_def_tramoseats(spec = c("TRfull", "TR0", "TR1", "TR2",
  "TR3", "TR4", "TR5"), estimate.from = NA_character_,
  estimate.to = NA_character_, estimate.first = NA_integer_,
  estimate.last = NA_integer_, estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_, estimate.tol = NA_integer_,
  estimate.eml = NA, estimate.urfinal = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.fct = NA_integer_, usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA, usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA, usrdef.varEnabled = NA, usrdef.var = NA,
  usrdef.varType = NA, usrdef.varCoef = NA, tradingdays.mauto = c(NA,
  "Unused", "FTest", "WaldTest"), tradingdays.pftd = NA_integer_,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined",
  "None"), tradingdays.leapyear = NA,
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA,
  "Separate_T", "Joint_F", "None"), easter.type = c(NA, "Unused",
  "Standard", "IncludeEaster", "IncludeEasterMonday"),
  easter.julian = NA, easter.duration = NA_integer_,
  easter.test = NA, outlier.enabled = NA,
  outlier.from = NA_character_, outlier.to = NA_character_,
  outlier.first = NA_integer_, outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_, outlier.exclLast = NA_integer_,
  outlier.ao = NA, outlier.tc = NA, outlier.ls = NA,
  outlier.so = NA, outlier.usedefcv = NA, outlier.cv = NA_integer_,
  outlier.eml = NA, outlier.tcrate = NA_integer_,
  automdl.enabled = NA, automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_, automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_, automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_, automdl.ljungboxlimit = NA_integer_,
  automdl.compare = NA, arima.mu = NA, arima.p = NA_integer_,
  arima.d = NA_integer_, arima.q = NA_integer_,
  arima.bp = NA_integer_, arima.bd = NA_integer_,
  arima.bq = NA_integer_, arima.coefEnabled = NA, arima.coef = NA,
  arima.coefType = NA, fcst.horizon = NA_integer_)

regarima_def_tramoseats(series, spec = c("TRfull", "TR0", "TR1", "TR2",
  "TR3", "TR4", "TR5"))

regarima_def_x13(series, spec = c("RG5c", "RG0", "RG1", "RG2c", "RG3",
  "RG4c"))

regarima_spec_def_x13(spec = c("RG5c", "RG0", "RG1", "RG2c", "RG3",
  "RG4c"), estimate.from = NA_character_, estimate.to = NA_character_,
  estimate.first = NA_integer_, estimate.last = NA_integer_,
  estimate.exclFirst = NA_integer_, estimate.exclLast = NA_integer_,
  estimate.tol = NA_integer_, transform.function = c(NA, "Auto",
  "None", "Log"), transform.adjust = c(NA, "None", "LeapYear",
  "LengthOfPeriod"), transform.aicdiff = NA_integer_,
  usrdef.outliersEnabled = NA, usrdef.outliersType = NA,
  usrdef.outliersDate = NA, usrdef.outliersCoef = NA,
  usrdef.varEnabled = NA, usrdef.var = NA, usrdef.varType = NA,
  usrdef.varCoef = NA, tradingdays.option = c(NA, "TradingDays",
  "WorkingDays", "UserDefined", "None"), tradingdays.autoadjust = NA,
  tradingdays.leapyear = c(NA, "LeapYear", "LengthOfPeriod", "None"),
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA, "Remove",
  "Add", "None"), easter.enabled = NA, easter.julian = NA,
  easter.duration = NA_integer_, easter.test = c(NA, "Add", "Remove",
  "None"), outlier.enabled = NA, outlier.from = NA_character_,
  outlier.to = NA_character_, outlier.first = NA_integer_,
  outlier.last = NA_integer_, outlier.exclFirst = NA_integer_,
  outlier.exclLast = NA_integer_, outlier.ao = NA, outlier.tc = NA,
  outlier.ls = NA, outlier.so = NA, outlier.usedefcv = NA,
  outlier.cv = NA_integer_, outlier.method = c(NA, "AddOne", "AddAll"),
  outlier.tcrate = NA_integer_, automdl.enabled = NA,
  automdl.acceptdefault = NA, automdl.cancel = NA_integer_,
  automdl.ub1 = NA_integer_, automdl.ub2 = NA_integer_,
  automdl.mixed = NA, automdl.balanced = NA,
  automdl.armalimit = NA_integer_, automdl.reducecv = NA_integer_,
  automdl.ljungboxlimit = NA_integer_, automdl.ubfinal = NA_integer_,
  arima.mu = NA, arima.p = NA_integer_, arima.d = NA_integer_,
  arima.q = NA_integer_, arima.bp = NA_integer_,
  arima.bd = NA_integer_, arima.bq = NA_integer_,
  arima.coefEnabled = NA, arima.coef = NA, arima.coefType = NA,
  fcst.horizon = NA_integer_)
}
\description{
Use \code{\link{x13_spec}} instead of \code{x13_spec_def}.

Use \code{\link{x13}} instead of \code{x13_def}.

Use \code{\link{tramoseats_spec}} instead of \code{tramoseats_spec_def}.

Use \code{\link{tramoseats}} instead of \code{tramoseats_def}.

Use \code{\link{regarima_spec_tramoseats}} instead of \code{regarima_spec_def_tramoseats}.

Use \code{\link{regarima_tramoseats}} instead of \code{regarima_def_tramoseats}.

Use \code{\link{regarima_x13}} instead of \code{regarima_def_x13}.

Use \code{\link{regarima_spec_x13}} instead of \code{regarima_spec_def_x13}.
}
