% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regarima_tests.R
\name{doornik_hansen_test}
\alias{doornik_hansen_test}
\title{Doornik-Hansen normality test}
\usage{
doornik_hansen_test(x)
}
\arguments{
\item{x}{object of class \code{"regarima"} or \code{"SA"}.}
}
\description{
Function compute the Doornik-Hansen normality test  on RegARIMA residuals
}
\examples{
mysa1 <- tramoseats_def(myseries, spec = "RSAfull")

# Actual tests :
mysa1$regarima$residuals.stat$tests

#Doornik Hansen test:
doornik_hansen_test(mysa1)
#Equivalent to:
doornik_hansen_test(mysa1$regarima)
}
