% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipi_c_eu.R
\docType{data}
\name{ipi_c_eu}
\alias{ipi_c_eu}
\title{Industrial Production Indices in manufacturing industry in the European Union}
\format{
A monthly \code{ts} object from January 1990 to December 2020 with 34 variables.
}
\source{
Eurostat, 'sts_inpr_m' database.
}
\usage{
ipi_c_eu
}
\description{
A dataset containing on monthly industrial production indices in manufacturing in the European Union (from sts_inpr_m dataset of Eurostat). Data are based 100 in 2015 and are unadjusted, i.e. neither seasonally adjusted nor calendar adjusted.
}
\details{
The dataset contains 34 time series corresponding to the following geographical area
\tabular{cl}{
BE        \tab Belgium                                         \cr
BG        \tab Bulgaria                                        \cr
CZ        \tab Czech Republic                                  \cr
DK        \tab Denmark                                         \cr
DE        \tab Germany (until 1990 former territory of the FRG)\cr
EE        \tab Estonia                                         \cr
IE        \tab Ireland                                         \cr
EL        \tab Greece                                          \cr
ES        \tab Spain                                           \cr
FR        \tab France                                          \cr
HR        \tab Croatia                                         \cr
IT        \tab Italy                                           \cr
CY        \tab Cyprus                                          \cr
LV        \tab Latvia                                          \cr
LT        \tab Lithuania                                       \cr
LU        \tab Luxembourg                                      \cr
HU        \tab Hungary                                         \cr
MT        \tab Malta                                           \cr
NL        \tab Netherlands                                     \cr
AT        \tab Austria                                         \cr
PL        \tab Poland                                          \cr
PT        \tab Portugal                                        \cr
RO        \tab Romania                                         \cr
SI        \tab Slovenia                                        \cr
SK        \tab Slovakia                                        \cr
FI        \tab Finland                                         \cr
SE        \tab Sweden                                          \cr
UK        \tab United Kingdom                                  \cr
NO        \tab Norway                                          \cr
CH        \tab Switzerland                                     \cr
ME        \tab Montenegro                                      \cr
MK        \tab Former Yugoslav Republic of Macedonia, the      \cr
RS        \tab Serbia                                          \cr
TR        \tab Turkey                                          \cr
BA        \tab Bosnia and Herzegovina
}
}
\keyword{datasets}
