% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_tramoseats.R
\name{jd_defTS}
\alias{jd_defTS}
\title{SA, TRAMO/SEATS}
\usage{
jd_defTS(series, spec = c("RSAfull", "RSA0", "RSA1", "RSA2", "RSA", "RSA4",
  "RSA5"), userdefined)
}
\arguments{
\item{series}{a univariate time series}

\item{spec}{model specification TRAMO/SEATS}

\item{userdefined}{vector with characters for additional output variables}
}
\value{
.
}
\description{
.
}
\details{
.
}
\examples{
mysa <- jd_defTS(myseries, "RSAfull")

}
\references{
Info on JDemtra+, usage and functions:
\url{https://ec.europa.eu/eurostat/cros/content/documentation_en}

BOX G.E.P. and JENKINS G.M. (1970), "Time Series Analysis: Forecasting and Control", Holden-Day, San Francisco.

BOX G.E.P., JENKINS G.M., REINSEL G.C. and LJUNG G.M. (2015), "Time Series Analysis: Forecasting and Control", John Wiley & Sons, Hoboken, N. J., 5th edition.
}
