% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_x13.R
\name{jd_defX13}
\alias{jd_defX13}
\title{SA, X13}
\usage{
jd_defX13(series, spec = c("RSA5c", "RSA0", "RSA1", "RSA2c", "RSA3", "RSA4c"))
}
\arguments{
\item{series}{a univariate time series}

\item{spec}{model specification}
}
\value{
.
}
\description{
.
}
\details{
.
}
\examples{
mysa <- jd_defX13(myseries, "RSA5c")

}
\references{
Info on JDemtra+, usage and functions:
\url{https://ec.europa.eu/eurostat/cros/content/documentation_en}

BOX G.E.P. and JENKINS G.M. (1970), "Time Series Analysis: Forecasting and Control", Holden-Day, San Francisco.

BOX G.E.P., JENKINS G.M., REINSEL G.C. and LJUNG G.M. (2015), "Time Series Analysis: Forecasting and Control", John Wiley & Sons, Hoboken, N. J., 5th edition.
}
