% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_regarima_specX13_s3.R
\name{jd_regarima_specX13}
\alias{jd_regarima_specX13}
\title{RegARIMA model specification, pre-adjustment in X13}
\usage{
jd_regarima_specX13(object = object, estimate.from = NA_character_,
  estimate.to = NA_character_, estimate.first = NA_integer_,
  estimate.last = NA_integer_, estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_, estimate.tol = NA_integer_,
  transform.function = c(NA_character_, "Auto", "None", "Log"),
  transform.adjust = c(NA_character_, "None", "LeapYear", "LengthOfPeriod"),
  transform.aicdiff = NA_integer_, usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA, usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA, usrdef.varEnabled = NA, usrdef.var = NA,
  usrdef.varType = NA, usrdef.varCoef = NA,
  tradingdays.option = c(NA_character_, "TradingDays", "WorkingDays", "None"),
  tradingdays.autoadjust = NA, tradingdays.leapyear = c(NA_character_,
  "LeapYear", "LengthOfPeriod", "None"), tradingdays.stocktd = NA_integer_,
  tradingdays.test = c(NA_character_, "Remove", "Add", "None"),
  easter.enabled = NA, easter.julian = NA, easter.duration = NA_integer_,
  easter.test = c(NA_character_, "Add", "Remove", "None"),
  outlier.enabled = NA, outlier.from = NA_character_,
  outlier.to = NA_character_, outlier.first = NA_integer_,
  outlier.last = NA_integer_, outlier.exclFirst = NA_integer_,
  outlier.exclLast = NA_integer_, outlier.ao = NA, outlier.tc = NA,
  outlier.ls = NA, outlier.so = NA, outlier.usedefcv = NA,
  outlier.cv = NA_integer_, outlier.method = c(NA_character_, "AddOne",
  "AddAll"), outlier.tcrate = NA_integer_, automdl.enabled = NA,
  automdl.acceptdefault = NA, automdl.cancel = NA_integer_,
  automdl.ub1 = NA_integer_, automdl.ub2 = NA_integer_,
  automdl.mixed = NA, automdl.balanced = NA,
  automdl.armalimit = NA_integer_, automdl.reducecv = NA_integer_,
  automdl.ljungboxlimit = NA_integer_, automdl.ubfinal = NA_integer_,
  arima.mu = NA, arima.p = NA_integer_, arima.d = NA_integer_,
  arima.q = NA_integer_, arima.bp = NA_integer_, arima.bd = NA_integer_,
  arima.bq = NA_integer_, arima.coefEnabled = NA, arima.coef = NA,
  arima.coefType = NA, fcst.horizon = NA_integer_)
}
\arguments{
\item{object}{object of class \code{c("JD_RegArima_Spec","X13")} or of class \code{c("JD_RegArima","X13")}.}

\item{estimate.from}{character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with \code{estimate.to}.}

\item{estimate.to}{character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with \code{estimate.from}.}

\item{estimate.first}{numeric specifying the number of periods considered at the beginning of the series.}

\item{estimate.last}{numeric specifying the number of periods considered at the end of the series.}

\item{estimate.exclFirst}{numeric specifying the number of periods excluded at the beginning of the series. Can be combined with \code{estimate.exclLast}.}

\item{estimate.exclLast}{numeric specifying the number of periods excluded at the end of the series. Can be combined with \code{estimate.exclFirst}.}

\item{estimate.tol}{numeric, convergence tolerance. The absolute changes in the log-likelihood function are compared to this value to check for the convergence of the estimation iterations.}

\item{transform.function}{transformation of the input series: \code{"None"} - no transformation of the series; \code{"Log"} - takes the log of the series; \code{"Auto"} - the program tests for the log-level specification.}

\item{transform.adjust}{pre-adjustment of the input series for length of period or leap year effects: \code{"None"} - no adjustment; \code{"LeapYear"} - leap year effect; \code{"LengthOfPeriod"} - length of period. Modifications of this variable are taken into account only when \code{transform.function} is set to \code{"Log"}.}

\item{transform.aicdiff}{numeric defining the difference in AICC needed to accept no transformation when the automatic transformation selection is chosen (considered only when \code{transform.function} is set to \code{"Auto"}).

Control variables for the pre-specified outliers. The pre-specified outliers are used in the model only if they are enabled (\code{usrdef.outliersEnabled=TRUE}) and the outliers' type (\code{usrdef.outliersType}) and date (\code{usrdef.outliersDate}) are provided.}

\item{usrdef.outliersEnabled}{logicals. If \code{TRUE} the program uses the pre-specified outliers.}

\item{usrdef.outliersType}{vector defining the outliers' type. Possible types are: \code{("AO")} - additive, \code{("LS")} - level shift, \code{("TC")} - transitory change, \code{("SO")} -  seasonal outlier. E.g.: \code{ usrdef.outliersType =c("AO","AO","LS")}.}

\item{usrdef.outliersDate}{vector defining the outliers' date. The dates should be characters in format "YYYY-MM-DD". E.g.: \code{usrdef.outliersDate=c("2009-10-01","2005-02-01","2003-04-01")}.}

\item{usrdef.outliersCoef}{vector providing fixed coefficients for the outliers. The coefficients can't be fixed if  \code{ transform.function} is set to \code{"Auto"} - the series transformation need to be pre-defined. E.g.: \code{ usrdef.outliersCoef=c(200,170,20)}.

Control variables for the user-defined variables:}

\item{usrdef.varEnabled}{logicals. If \code{TRUE} the program uses the user-defined variables.}

\item{usrdef.var}{time series (\code{ts}) or matrix of time series (\code{mts}) with the user-defined variables.}

\item{usrdef.varType}{vector of character(s) defining the user-defined variables component type. Possible types are: \code{"Undefined", "Series", "Trend", "Seasonal", "SeasonallyAdjusted", "Irregular"}. If not specified, the program will assign the \code{"Undefined"} type.}

\item{usrdef.varCoef}{vector providing fixed coefficients for the user-defined variables. The coefficients can't be fixed if  \code{ transform.function} is set to \code{"Auto"} - the series transformation need to be pre-defined.}

\item{tradingdays.option}{defines the type of the trading days regression variables: \code{"TradingDays"} -  six day-of-the-week regression variables; \code{"WorkingDays"} - one working/non-working day contrast variable; \code{"None"} - no correction for trading days and working days effects. \code{"None"} has also to be chosen for the "day-of-week effects" correction (\code{tradingdays.stocktd} to be modified accordingly).}

\item{tradingdays.autoadjust}{logicals. If \code{TRUE} the program corrects automatically for the leap year effect. Modifications of this variable are taken into account only when \code{transform.function} is set to \code{"Auto"}.}

\item{tradingdays.leapyear}{option for including the leap-year effect in the model: \code{"LeapYear"} - leap year effect; \code{"LengthOfPeriod"} - length of period, \code{"None"} - no effect included. The leap-year effect can be pre-specified in the model only if the input series was not pre-adjusted (\code{transform.adjust} set to \code{"None"}) and the automatic correction for the leap-year effect was not selected (\code{tradingdays.autoadjust} set to \code{FALSE}).}

\item{tradingdays.stocktd}{numeric indicating the day of the month when inventories and other stock are reported (to denote the last day of the month set the variable to 31). Modifications of this variable are taken into account only when \code{tradingdays.option} is set to \code{"None"}.}

\item{tradingdays.test}{defines the pre-tests for the significance of the trading day regression variables based on the AICC statistics: \code{"Add"} - the trading day variables are not included in the initial regression model but can be added to the RegARIMA model after the test; \code{"Remove"} - the trading day variables belong to the initial regression model but can be removed from the RegARIMA model after the test; \code{"None"} - the trading day variables are not pre-tested and are included in the model.}

\item{easter.enabled}{logicals. If \code{TRUE} the program considers the Easter effect in the model.}

\item{easter.julian}{logicals. If \code{TRUE} the program uses the Julian Easter (expressed in Gregorian calendar).}

\item{easter.duration}{numeric indicating the duration of the Easter effect (length in days, between 1 and 20).}

\item{easter.test}{defines the pre-tests for the significance of the Easter effect based on the t-statistic (Easter effect is considered as significant if the t-statistic is greater than 1.96) : \code{"Add"} - the Easter effect variable is not included in the initial regression model but can be added to the RegARIMA model after the test; \code{"Remove"} - the Easter effect variable belong to the initial regression model but can be removed from the RegARIMA model after the test; \code{"None"} - the Easter effect variable is not pre-tested and is included in the model.}

\item{outlier.enabled}{logicals. If \code{TRUE} the automatic detection of outliers is enabled in the defined time span.

The time span of the series to be searched for outliers (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: \code{outlier.from, outlier.to, outlier.first, outlier.last, outlier.exclFirst} and \code{outlier.exclLast}; where \code{outlier.from} and \code{outlier.to} have priority over remaining span control variables, \code{outlier.last} and \code{outlier.first} have priority over \code{outlier.exclFirst} and \code{outlier.exclLast}, and \code{outlier.last} has priority over \code{outlier.first}.}

\item{outlier.from}{character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with \code{outlier.to}.}

\item{outlier.to}{character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with \code{outlier.from}.}

\item{outlier.first}{numeric specifying the number of periods considered at the beginning of the series.}

\item{outlier.last}{numeric specifying the number of periods considered at the end of the series.}

\item{outlier.exclFirst}{numeric specifying the number of periods excluded at the beginning of the series. Can be combined with \code{outlier.exclLast}.}

\item{outlier.exclLast}{numeric specifying the number of periods excluded at the end of the series. Can be combined with \code{outlier.exclFirst}.}

\item{outlier.ao}{logicals. If \code{TRUE} the automatic detection of additive outliers is enabled (\code{outlier.enabled} must be also set to \code{TRUE}).}

\item{outlier.tc}{logicals. If \code{TRUE} the automatic detection of transitory changes is enabled (\code{outlier.enabled} must be also set to \code{TRUE}).}

\item{outlier.ls}{logicals. If \code{TRUE} the automatic detection of level shifts is enabled (\code{outlier.enabled} must be also set to \code{TRUE}).}

\item{outlier.so}{logicals. If \code{TRUE} the automatic detection of seasonal outliers is enabled (\code{outlier.enabled} must be also set to \code{TRUE}).}

\item{outlier.usedefcv}{logicals. If \code{TRUE} the critical value for the outliers' detection procedure is automatically determined by the number of observations in the outlier detection time span. If \code{FALSE} the procedure uses the inputted critical value (\code{outlier.cv}).}

\item{outlier.cv}{numeric. Inputted critical value for the outliers' detection procedure. The modification of this variable is taken into account only when \code{outlier.usedefcv} is set to \code{FALSE}.}

\item{outlier.method}{determines how the program successively adds detected outliers to the model. At present only the \code{AddOne} method is supported.}

\item{outlier.tcrate}{numeric. The rate of decay for the transitory change outlier.}

\item{automdl.enabled}{logicals. If \code{TRUE} the automatic modelling of the ARIMA model is enabled. If \code{FALSE} the parameters of the ARIMA model can be specified.

Control variables for the automatic modelling of the ARIMA model (\code{automdl.enabled} is set to \code{TRUE}):}

\item{automdl.acceptdefault}{logicals. If \code{TRUE} the default model (ARIMA(0,1,1)(0,1,1)) may be chosen in the first step of the automatic model identification. If the Ljung-Box Q statistics for the residuals is acceptable, the default model is accepted and no further attempt will be made to identify any other.}

\item{automdl.cancel}{numeric, cancelation limit. If the difference in moduli of an AR and an MA roots (when estimating ARIMA(1,0,1)(1,0,1) models in the second step of the automatic identification of the differencing orders) is smaller than cancelation limit, the two roots are assumed equal and cancel out.}

\item{automdl.ub1}{numeric, first unit root limit. It is the threshold value for the initial unit root test in the automatic differencing procedure. When one of the roots in the estimation of the ARIMA(2,0,0)(1,0,0) plus mean model, performed in the first step of the automatic model identification procedure, is larger than first unit root limit in modulus, it is set equal to unity.}

\item{automdl.ub2}{numeric, second unit root limit. When one of the roots in the estimation of the ARIMA(1,0,1)(1,0,1) plus mean model, which is performed in the second step of the automatic model identification procedure, is larger than second unit root limit in modulus, it is checked if there is a common factor in the corresponding AR and MA polynomials of the ARMA model that can be cancelled (see \code{automdl.cancel}). If there is no cancellation, the AR root is set equal to unity (i.e. the differencing order changes).}

\item{automdl.mixed}{logicals. The variable controls whether ARIMA models with non-seasonal AR and MA terms or seasonal AR and MA terms will be considered in the automatic model identification procedure. If \code{FALSE} a model with AR and MA terms in both the seasonal and non-seasonal parts of the model can be acceptable, provided there are not AR and MA terms in either the seasonal or non-seasonal.}

\item{automdl.balanced}{logicals. If \code{TRUE}, the automatic model identification procedure will have a preference for balanced models (i.e. models for which the order of the combined AR and differencing operator is equal to the order of the combined MA operator).}

\item{automdl.armalimit}{numeric, arma limit. It is the threshold value for t-statistics of ARMA coefficients and constant term used for the final test of model parsimony. If the highest order ARMA coefficient has a t-value less than this value in magnitude, the order of the model is reduced. Also if the constant term has a t-value less than arma limit in magnitude, it is removed from the set of regressors.}

\item{automdl.reducecv}{numeric, ReduceCV.  The percentage by which the outlier's critical value will be reduced when an identified model is found to have a Ljung-Box statistic with an unacceptable confidence coefficient. The parameter should be between 0 and 1, and will only be active when automatic outlier identification is enabled. The reduced critical value will be set to (1-ReduceCV)xCV, where CV is the original critical value.}

\item{automdl.ljungboxlimit}{numeric, Ljung Box limit. Acceptance criterion for the confidence intervals of the Ljung-Box Q statistic. If the LjungBox Q statistics for the residuals of a final model is greater than Ljung Box limit, the model is rejected, the outlier critical value is reduced, and model and outlier identification (if specified) is redone with a reduced value.}

\item{automdl.ubfinal}{numeric, final unit root limit. The threshold value for the final unit root test. If the magnitude of an AR root for the final model is less than the final unit root limit, a unit root is assumed, the order of the AR polynomial is reduced by one, and the appropriate order of the differencing (non-seasonal, seasonal) is increased. The parameter value should be greater than one.

Control variables for the non-automatic modelling of the ARIMA model (\code{automdl.enabled} is set to \code{FALSE}):}

\item{arima.mu}{logicals. If \code{TRUE}, the mean is considered as part of the ARIMA model.}

\item{arima.p}{numeric. The order of the non-seasonal autoregressive (AR) polynomial.}

\item{arima.d}{numeric. Regular differencing order.}

\item{arima.q}{numeric. The order of the non-seasonal moving average (MA) polynomial.}

\item{arima.bp}{numeric. The order of the seasonal autoregressive (AR) polynomial.}

\item{arima.bd}{numeric. Seasonal differencing order.}

\item{arima.bq}{numeric. The order of the seasonal moving average (MA) polynomial.

Control variables for the user-defined ARMA coefficients. Coefficients can be defined for the regular and seasonal autoregressive (AR) polynomials and moving average (MA) polynomials. The model considers the coefficients only if the procedure for their estimation (\code{arima.coefType}) is provided, and the number of provided coefficients matches the sum of (regular and seasonal) AR and MA orders (\code{p,q,bp,bq}).}

\item{arima.coefEnabled}{logicals. If \code{TRUE} the program uses the user-defined ARMA coefficients.}

\item{arima.coef}{vector providing the coefficients for the regular and seasonal AR and MA polynominals. The length of the vector must equal the sum of the regular and seasonal AR and MA orders. The coefficients shall be provided in the order: regular AR (\emph{Phi} - \code{p} elements), regular MA  (\emph{Theta} - \code{q} elements), seasonal AR (\emph{BPhi} - \code{bp} elements) and seasonal MA (\emph{BTheta} - \code{bq} elements). E.g.: \code{arima.coef=c(0.6,0.7)} with \code{arima.p=1, arima.q=0,arima.bp=1} and \code{arima.bq=0}.}

\item{arima.coefType}{vector defining ARMA coefficients estimation procedure. Possible procedures are: \code{"Undefined"} - no use of user-defined input (i.e. coefficients are estimated), \code{"Fixed"} - fixes the coefficients at the value provided by the user, \code{"Initial"} - the value defined by the user is used as initial condition. For orders for which the coefficients shall not be defined, the \code{arima.coef} can be set to \code{NA} or \code{0} or the \code{arima.coefType} can be set to \code{"Undefined"}.
E.g.: \code{arima.coef = c(-0.8,-0.6,NA)}, \code{arima.coefType = c("Fixed","Fixed","Undefined")}.}

\item{fcst.horizon}{numeric, forecasts horizon. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values). By default the program generates two years forecasts (\code{fcst.horizon} set to \code{-2}).}
}
\value{
A list of class \code{c("JD_RegArima_Spec","X13")} with the below components. Each component refers to different part of the RegARIMA model specification, mirroring the arguments of the function (for details see arguments description). Each of the lowest-level component (except span, pre-specified outliers, user-defined variables and pre-specified ARMA coefficients) is structured within a data frame with columns denoting different variables of the model specification and rows referring to: first row - base specification, specification of the \code{c("JD_RegArima","X13")} or \code{c("JD_RegArima_Spec","X13")} object as provided within the argument \code{object}; second row - user modifications as specified by the remaining arguments of the function (e.g.: \code{arima.d}); and third row - final model specification, values that will be used in the function \code{\link{jd_regarima}}. The final specification (third row) shall include user modifications (row two) unless they were wrongly specified. The pre-specified outliers, user-defined variables and pre-specified ARMA coefficients consist of a list with the \code{Predefined} (base model specification) and \code{Final} values.

\item{estimate}{data frame. Variables referring to: \code{span} - time span for the model estimation, \code{tolerance} - argument \code{estimate.tol}. The final values can be also accessed with the function \code{\link{s_estimate}}.}

\item{transform}{data frame. Variables referring to: \code{tfunction} - argument \code{transform.function}, \code{adjust} - argument \code{transform.adjust}, \code{aicdiff} - argument \code{transform.aicdiff}. The final values can be also accessed with the function \code{\link{s_transform}}.}

\item{regression}{list including the information on the user-defined variables (\code{userdef}), \code{trading.days} effect and \code{easter} effect. The user-defined part includes: \code{specification} - data frame with the information if pre-specified outliers (\code{outlier}) and user-defined variables (\code{variables}) are included in the model and if fixed coefficients are used (\code{outlier.coef} and \code{variables.coef}). The final values can be also accessed with the function \code{\link{s_usrdef}}; \code{outliers} - matrixes with the outliers (\code{Predefined} and \code{Final}). The final outliers can be also accessed with the function \code{\link{s_preOut}}; and \code{variables} - list with the \code{Predefined} and \code{Final} user-defined variables (\code{series}) and its description (\code{description}) including the information on the variable type and values of fixed coefficients. The final user-defined variables can be also accessed with the function \code{\link{s_preVar}}.
Within the data frame \code{trading.days} variables refer to: \code{option} - argument \code{tradingdays.option, autoadjust} - argument \code{tradingdays.autoadjust, leapyear} - argument \code{tradingdays.leapyear, stocktd} - argument \code{tradingdays.stocktd, test} - argument \code{tradingdays.test}. The final \code{trading.days} values can be also accessed with the function \code{\link{s_td}}. Within the data frame \code{easter} variables refer to: \code{enabled} - argument \code{easter.enabled, julian} - argument \code{easter.julian, duration} - argument \code{easter.duration, test} - argument \code{easter.test}. The final \code{easter} values can be also accessed with the function \code{\link{s_easter}}.}

\item{outliers}{data frame. Variables referring to: \code{enabled} - argument \code{outlier.enabled},  \code{span}  - time span for the outliers' detection, \code{ao} - argument \code{outlier.ao, tc} - argument \code{outlier.tc, ls} - argument \code{outlier.ls, so} - argument \code{outlier.so, usedefcv} - argument \code{outlier.usedefcv, cv} - argument \code{outlier.cv, method} - argument \code{outlier.method, tcrate} - argument \code{outlier.tcrate}. The final values can be also accessed with the function \code{\link{s_out}}.}

\item{arima}{list containing a data frame with the ARIMA settings (\code{specification}) and matrixes with the information on the pre-specified ARMA coefficients (\code{coefficients}). The matrix \code{Predefined} refers to the pre-defined model specification and matrix \code{Final} to the final specification. Both matrixes contain the value of the ARMA coefficients and the procedure for its estimation.
Within the data frame \code{specification} the variable \code{enabled} refer to the argument \code{automdl.enabled} and all the remaining variables (\code{automdl.acceptdefault, automdl.cancel, automdl.ub1, automdl.ub2, automdl.mixed, automdl.balanced, automdl.armalimit, automdl.reducecv, automdl.ljungboxlimit, automdl.ubfinal, arima.mu, arima.p, arima.d, arima.q, arima.bp, arima.bd, arima.bq}) to the respective function arguments.
The final values of the \code{specification} can be also accessed with the function \code{\link{s_arima}} and final pre-specified ARMA coefficients with the function \code{\link{s_arimaCoef}}.}

\item{forecast}{data frame with the forecast horizon (argument \code{fcst.horizon}). The final value can be also accessed with the function \code{\link{s_fcst}}.}


\item{span}{matrix containing the final time span for the model estimation and outliers' detection. Contains the same information as the variable span in the data frames estimate and outliers. The matrix can be also accessed with the function \code{\link{s_span}}.}
}
\description{
\code{jd_regarima_specX13} creates (and/or modifies) a \code{c("JD_RegArima_Spec","X13")} class object with the RegARIMA model specification for the X13 method. The object is created from a \code{c("JD_RegArima","X13")} or \code{c("JD_RegArima_Spec","X13")} class object.
}
\examples{

  myreg <- jd_regarima_defX13(myseries, spec=c("RG5c"))
  myspec<- jd_regarima_specX13(myreg)
  myreg1<- jd_regarima(myseries,myspec)

# Modify the model specification from a "JD_RegArima" object
  myspec2 <- jd_regarima_specX13(myreg,tradingdays.option = "WorkingDays")
  myreg2 <- jd_regarima(myseries,myspec2)

# Modify the model specification from a "JD_RegArima_Spec" object
  myspec3 <- jd_regarima_specX13(myspec,tradingdays.option = "WorkingDays")
  myreg3 <- jd_regarima(myseries,myspec3)

# Pre-specified outliers
  myreg <- jd_regarima_defX13(myseries, spec=c("RG5c"))
  myspec1<-jd_regarima_specX13(myreg, usrdef.outliersEnabled = TRUE,
                             usrdef.outliersType = c("LS","AO"),
                             usrdef.outliersDate=c("2008-10-01","2002-01-01"),
                             usrdef.outliersCoef = c(36000,14000),
                             transform.function = "None")

  myreg1 <- jd_regarima(myseries, myspec1)
  myreg1
  s_preOut(myreg1)

# User-defined variables
  myreg <- jd_regarima_defX13(myseries, spec=c("RG5c"))
  var1 <- ts(rnorm(length(myseries))*10,start = c(2001, 12), frequency = 12)
  var2 <- ts(rnorm(length(myseries))*100,start = c(2001, 12), frequency = 12)
  var <-ts(matrix(c(var1,var2), ncol=2),start = c(2001, 12), frequency = 12)

  myspec1 <- jd_regarima_specX13(myreg, usrdef.varEnabled = TRUE,
                                  usrdef.var = var)
  myreg1 <- jd_regarima(myseries,myspec1)
  myreg1

  myspec2 <- jd_regarima_specX13(myreg, usrdef.varEnabled = TRUE,
                                  usrdef.var = var1, usrdef.varCoef = c(2),
                                  transform.function = "None")
  myreg2 <- jd_regarima(myseries, myspec2)
  s_preVar(myreg2)
  myreg2

# Pre-specified ARMA coefficients
  myreg <- jd_regarima_defX13(myseries, spec=c("RG5c"))
  myspec1 <- jd_regarima_specX13(myreg, automdl.enabled =FALSE,
                              arima.p=1,arima.q=1, arima.bp=0, arima.bq=1,
                              arima.coefEnabled = TRUE,
                              arima.coef = c(-0.8,-0.6,0),
                              arima.coefType = c(rep("Fixed",2),"Undefined"))

  s_arimaCoef(myspec1)
  myreg1 <- jd_regarima(myseries, myspec1)
  myreg1
}
\references{
Info on JDemtra+, usage and functions:
\url{https://ec.europa.eu/eurostat/cros/content/documentation_en}
}
