% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsa2r.R
\name{jSA}
\alias{jSA}
\alias{get_dictionary}
\alias{get_indicators}
\title{Functions around 'jSA' objects}
\usage{
get_dictionary(x)

get_indicators(x, ...)
}
\arguments{
\item{x}{a \code{"jSA"} object.}

\item{...}{characters containing the names of the indicators to extract.}
}
\value{
\code{get_dictionary} a vector of characters and \code{get_indicators} returns a list containing the indicators that are extracted.
}
\description{
\code{get_dictionary} returns the indicators that can be extracted from \code{jSA} and \code{get_indicators} returns a list of indicators.
\code{jSA} constructs a object of class \code{jSA}.
}
\details{
A \code{jSA} object is a list with three elements: \itemize{
\item \code{"result"}: the Java object with the results of a seasonal adjustment or a pre-adjustment method.
\item \code{"spec"}: the Java object with the specification of a seasonal adjustment or a pre-adjustment method.
\item \code{"dictionary"}: the Java object with dictionnary of a seasonal adjustment or a pre-adjustment method. In particular, it contains all the user-defined regressors.
 }

\code{get_dictionary} returns the list of indicators that can be extracted from a \code{jSA} object by the function \code{get_indicators}.
}
\examples{
myseries <- ipi_c_eu[, "FR"]
mysa <- jx13(myseries, spec = "RSA5c")
get_dictionary(mysa)

get_indicators(mysa, "decomposition.b2", "decomposition.d10")
}
