\name{myseries}
\alias{myseries}
\docType{data}
\title{
    dataset myseries
}
\description{
    timeseries for regarima test
}
\format{
    ts series with 192 obeservations
}
\source{
    optional: reference to a publication or URL from which the data were obtained
}

\examples{
data(myseries)
str(myseries)
}
