% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_regarima_plot_s3.R, R/jd_sa_plot.R
\name{plot}
\alias{plot}
\alias{plot.JD_RegArima}
\alias{plot}
\alias{plot.Decomp_X13}
\alias{plot}
\alias{plot.Decomp_TS}
\alias{plot}
\alias{plot.final}
\alias{plot}
\alias{plot.SA}
\title{Plotting decomposition or final results of a seasonal adjustment}
\usage{
\method{plot}{JD_RegArima}(x, which = c(1:6), dec_zoom = FALSE,
  caption = list("Residuals", "Histogram of residuals", "Normal Q-Q",
  "ACF of residuals", "PACF of residuals", "Decomposition"),
  sub.caption = NULL, main = "", ask = prod(par("mfcol")) < length(which)
  && dev.interactive(), ..., qqline = TRUE, cex.caption = 1,
  cex.oma.main = 1.25)

\method{plot}{Decomp_X13}(x, first_date, last_date, ...)

\method{plot}{Decomp_TS}(x, first_date, last_date, ...)

\method{plot}{final}(x, first_date, last_date, forecast = TRUE,
  type_chart = c("sa-trend", "cal-seas-irr"), ask = length(type_chart) > 1
  && dev.interactive(), ...)

\method{plot}{SA}(x, ...)
}
\arguments{
\item{x}{the object to plot.}

\item{which}{XXXXXX}

\item{dec_zoom}{XXXXXX}

\item{caption}{XXXXXX}

\item{sub.caption}{XXXXXX}

\item{main}{main title.}

\item{ask}{logical value. If TRUE, the user will in future be prompted before a new graphical page is started.}

\item{...}{other parameters.}

\item{qqline}{XXXXXX}

\item{cex.caption}{XXXXXX}

\item{cex.oma.main}{XXXXXX}

\item{first_date}{the first date to start the plot. If missing the plot starts at the beginning of the time-series.}

\item{last_date}{the last date to end the plot. If missing the plot ends at the end of the time-series (eventually, including forecast).}

\item{forecast}{logical indicating if forecasts should be included in the plot.}

\item{type_chart}{character vector indicating which type of chart to plot.}
}
\description{
Plotting methods for objects around the seasonal adjustment : "JD_RegArima" for regARIMA,"Decomp_X13" and "Decomp_TS" for the decomposition with X13 and TRAMO-SEATS, "final" for the final results and "SA" for the entire seasonal adjusment object. The function \code{plot.SA} just calls the function \code{plot.final}.
}
