% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_spec.R
\name{s_seats}
\alias{s_seats}
\title{Access model specification, SA/TRAMO-SEATS}
\usage{
s_seats(object = NA)
}
\arguments{
\item{object}{object of class: \code{c("SA","TRAMO_SEATS")} or \code{c("SA_Spec","TRAMO_SEATS")}.}
}
\value{
A data.frame with the \emph{seats} variables. For details see: \emph{Value} of the \code{c("SA_Spec","TRAMO_SEATS")} class object.
}
\description{
\code{s_seats} access the \code{seats} part of the model specification.
}
\examples{
  myspec1 <-jd_tramoseats_specDef(spec=c("RSAfull"),seats.approx ="Noisy")

  s_seats(myspec1)

}
\references{
Info on JDemtra+, usage and functions:
\url{https://ec.europa.eu/eurostat/cros/content/documentation_en}
}
