% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd_spec.R
\name{s_x11}
\alias{s_x11}
\title{Access model specification, SA/X13}
\usage{
s_x11(object = NA)
}
\arguments{
\item{object}{object of class: \code{c("SA","X13")} or \code{c("SA_Spec","X13")}.}
}
\value{
A data.frame with the \emph{seats} variables. For details see: \emph{Value} of the \code{c("SA_Spec","X13")} class object.
}
\description{
\code{s_x11} access the \code{x11} part of the model specification.
}
\examples{
  myspec1 <-jd_x13_specDef(spec=c("RSA5c"),x11.seasonalComp = FALSE)

  s_x11(myspec1)

}
\references{
Info on JDemtra+, usage and functions:
\url{https://ec.europa.eu/eurostat/cros/content/documentation_en}
}
