% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSpec.R
\name{saveSpec}
\alias{saveSpec}
\alias{saveSpec}
\alias{loadSpec}
\title{Saving and loading a model specification, SA and pre-adjustment in X13 and TRAMO-SEATS}
\usage{
saveSpec(object, file = "spec.RData")

loadSpec(file = "spec.RData")
}
\arguments{
\item{object}{object of one of the classes: \code{c("SA_spec","X13")}, \code{c("SA_spec","TRAMO_SEATS")}, \code{c("SA","X13")}, \code{c("SA","TRAMO_SEATS")}, \code{c("regarima_spec","X13")}, \code{c("regarima_spec","TRAMO_SEATS")}, \code{c("regarima","X13")}, \code{c("regarima","TRAMO_SEATS")}.}

\item{file}{(path and) name of the file where the model specification will be saved (have been saved).}
}
\value{
\code{loadSpec}  returns an object of class \code{"SA_spec"} or \code{"regarima_spec"}.
}
\description{
\code{saveSpec} saves a SA or RegARIMA model specification. \code{loadSpec} loads the previously saved model specification.
}
\details{
\code{saveSpec} saves the final model specification of a \code{"SA_spec"}, \code{"SA"}, \code{"regarima_spec"} or \code{"regarima"} class object.
\code{loadSpec} loads the previously saved model specification. It creates a \code{c("SA_spec","X13")}, \code{c("sA_spec","TRAMO_SEATS")}, \code{c("regarima_spec","X13")} or \code{c("regarima_spec","TRAMO_SEATS")} class object, in line with the class of the previously saved model specification.
}
\examples{
  myreg1 <-regarimaDefX13(myseries, spec=c("RG5c"))
  myspec2 <-regarima_specX13(myreg1, estimate.from = "2005-10-01", outlier.from = "2010-03-01")
  myreg2 <-regarima(myseries,myspec2)

  myreg3 <-regarimaDefTS(myseries, spec=c("TRfull"))
  myspec4 <-regarima_specTS(myreg3,tradingdays.mauto ="Unused",tradingdays.option ="WorkingDays",
                               easter.type = "Standard", automdl.enabled = FALSE, arima.mu = TRUE)
  myreg4 <-regarima(myseries,myspec4)

  myspec6 <- x13_specDef("RSA5c")
  mysa6 <- x13(myseries,myspec6)

  myspec7 <- tramoseats_specDef("RSAfull")
  mysa7 <- tramoseats(myseries,myspec7)

  # Save the model specification from a c("regarima_spec","X13") class object
  saveSpec(myspec2,"specx13.RData")
  # Save the model specification from a c("regarima","X13") class object
  saveSpec(myreg2,"regx13.RData")
  # Save the model specification from a c("regarima_spec","TRAMO_SEATS") class object
  saveSpec(myspec4,"specTS.RData")
  # Save the model specification from a c("regarima","TRAMO_SEATS") class object
  saveSpec(myreg4,"regTS.RData")
  # Save model from a c("SA_spec","X13") class object
  saveSpec(myspec6,"specFullx13.RData")
  # Save model from a c("SA","X13") class object
  saveSpec(mysa6,"sax13.RData")
  # Save model from a c("SA_spec","TRAMO_SEATS") class object
  saveSpec(myspec7,"specFullTS.RData")
  # Save model from a c("SA","TRAMO_SEATS") class object
  saveSpec(mysa7,"saTS.RData")

  # Load the model specification
  myspec2a <- loadSpec("specx13.RData")
  myspec2b <- loadSpec("regx13.RData")
  myspec4a <- loadSpec("specTS.RData")
  myspec4b <- loadSpec("regTS.RData")
  myspec6a <- loadSpec("specFullx13.RData")
  myspec6b <- loadSpec("sax13.RData")
  myspec7a <- loadSpec("specFullTS.RData")
  myspec7b <- loadSpec("saTS.RData")

  regarima(myseries,myspec2a)
  regarima(myseries,myspec4a)
  x13(myseries,myspec6a)
  x13(myseries,myspec6b)
  tramoseats(myseries,myspec7a)
  tramoseats(myseries,myspec7b)

}
\references{
Info on JDemtra+, usage and functions:
\url{https://ec.europa.eu/eurostat/cros/content/documentation_en}
}
