% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_defined.R
\name{user_defined_variables}
\alias{user_defined_variables}
\title{Get the names of the user-defined variables}
\usage{
user_defined_variables(sa_object = c("X13-ARIMA", "TRAMO-SEATS"))
}
\arguments{
\item{sa_object}{a character: \code{"X13-ARIMA"} to get the additional output variables available for the X13-ARIMA method and \code{"TRAMO-SEATS"} for the TRAMO-SEATS method.}
}
\description{
Function to get the names of the additional output variables that can be defined in
\code{\link{x13}} and \code{\link{tramoseats}} with the parameter \code{userdefined}.
}
\examples{
user_defined_variables("X13-ARIMA")
user_defined_variables("TRAMO-SEATS")
}
