% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_workspace.R
\name{new_workspace}
\alias{new_workspace}
\alias{new_workspace}
\alias{new_multiprocessing}
\title{Create a workspace or a multi-processing}
\usage{
new_workspace()

new_multiprocessing(workspace, name)
}
\arguments{
\item{workspace}{a workspace object}

\item{name}{character name of the new multiprocessing}
}
\value{
\code{new_workspace()} returns an object of class \code{workspace} and
\code{new_multiprocessing()} returns an object of class \code{multiprocessing}.
}
\description{
Functions to create a JDemetra+ workspace (\code{new_workspace()})
add a multi-processing to it (\code{new_multiprocessing()}).
}
\examples{
\dontrun{
# Create and export a empty JDemetra+ workspace
wk <- new_workspace()
mp <- new_multiprocessing(wk, "sa1")
}

}
\seealso{
\code{\link{load_workspace}}, \code{\link{save_workspace}},
\code{\link{add_sa_item}}
}
