% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regarima_plot.R, R/sa_plot.R
\name{plot}
\alias{plot}
\alias{plot.regarima}
\alias{plot.decomposition_X11}
\alias{plot.decomposition_SEATS}
\alias{plot.final}
\alias{plot.SA}
\title{Plotting regarima, decomposition or final results of SA}
\usage{
\method{plot}{regarima}(x, which = c(1:6), caption = list("Residuals",
  "Histogram of residuals", "Normal Q-Q", "ACF of residuals",
  "PACF of residuals", "Decomposition"), ask = prod(par("mfcol")) <
  length(which) && dev.interactive(), ...)

\method{plot}{decomposition_X11}(x, first_date, last_date, ...)

\method{plot}{decomposition_SEATS}(x, first_date, last_date, ...)

\method{plot}{final}(x, first_date, last_date, forecast = TRUE,
  type_chart = c("sa-trend", "cal-seas-irr"), ask = length(type_chart)
  > 1 && dev.interactive(), ...)

\method{plot}{SA}(x, ...)
}
\arguments{
\item{x}{the object to plot.}

\item{which}{vector with numerics specifying which graphs should be plotted: (1) "Residuals", (2) "Histogram of residuals", (3) "Normal Q-Q", (4) "ACF of residuals", (5) "PACF of residuals", (6) "Decomposition", (7) "Decomposition - zoom".}

\item{caption}{list with the graphs titles.}

\item{ask}{logicals. If \code{TRUE}, the user will in future be prompted before a new graphical page is started.}

\item{...}{other parameters.}

\item{first_date}{the first date to start the plot. If missing the plot starts at the beginning of the time-series.}

\item{last_date}{the last date to end the plot. If missing the plot ends at the end of the time-series (eventually, including forecast).}

\item{forecast}{logical indicating if forecasts should be included in the plot. If \code{TRUE} the forecast is plotted.}

\item{type_chart}{character vector indicating which type of chart to plot.}
}
\description{
Plotting methods for the S3 class objects around the seasonal adjustment: \code{"regarima"} for RegARIMA,\code{"decomposition_X11"} and \code{"decomposition_SEATS"} for the decomposition with X13 and TRAMO-SEATS, \code{"final"} for the final SA results and \code{"SA"} for the entire seasonal adjusment object. The function \code{plot.SA} just calls the function \code{plot.final}.
}
