% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13.R
\name{x13}
\alias{x13}
\title{SA, X13}
\usage{
x13(series, spec, userdefined)
}
\arguments{
\item{series}{a univariate time series}

\item{spec}{user-defined model specification}

\item{userdefined}{vector with characters for additional output variables}
}
\value{
.
}
\description{
.
}
\details{
.
}
\examples{
mysa <- myseries

}
\references{
Info on JDemtra+, usage and functions:
\url{https://ec.europa.eu/eurostat/cros/content/documentation_en}
BOX G.E.P. and JENKINS G.M. (1970), "Time Series Analysis: Forecasting and Control", Holden-Day, San Francisco.

BOX G.E.P., JENKINS G.M., REINSEL G.C. and LJUNG G.M. (2015), "Time Series Analysis: Forecasting and Control", John Wiley & Sons, Hoboken, N. J., 5th edition.
}
